Flying fonts
Mini font is a simple class to make flying fonts. 
You hit a monster and want to display the damage, This would help you...
It works nicely, I've checked it.
Can display any flying text, It's not dynamic, so it will always will have the same direction and for the same time. 
You can simply adjust line 27 and 32 to adjust the direction and timing.